+++ ICD.Connect.Displays/Devices/AbstractDisplay.cs
@@ -193,7 +193,7 @@ namespace ICD.Connect.Displays.Devices
 		[PublicAPI]
 		public void SetPort(ISerialPort port)
 		{
-			m_ConnectionStateManager.SetPort(port);
+			m_ConnectionStateManager.SetPort(port, false);
 		}
 
 		/// <summary>
@@ -526,6 +526,17 @@ namespace ICD.Connect.Displays.Devices
 			UpdateCachedOnlineStatus();
 
 			Trust = settings.Trust;
+		}
+
+		/// <summary>
+		/// Override to add actions on StartSettings
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		protected override void StartSettingsFinal()
+		{
+			base.StartSettingsFinal();
+
+			m_ConnectionStateManager.Start();
 		}
 
 		#endregion