+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+ - Implemented StartSettings() to start communications with devices
+ - Cleaned up AbstractProjectorScreenDevice to use new LifecycleState
 
 ## [14.0.1] - 2020-06-16
 ### Changed